add Procfile

+++ src/main/resources/db/migration/V1__Init_DB.sql
@@ -1,21 +1,21 @@
 create sequence hibernate_sequence start 1 increment 1;
 
 create table message (
-    id int4 not null,
+    id int8 not null,
     filename varchar(255),
     tag varchar(255),
     text varchar(2048) not null,
-    user_id int4,
+    user_id int8,
     primary key (id)
 );
 
 create table user_role (
-    user_id int4 not null,
+    user_id int8 not null,
     roles varchar(255)
 );
 
 create table usr (
-    id int4 not null,
+    id int8 not null,
     activation_code varchar(255),
     active boolean not null,
     email varchar(255),
